ci(4735): satisfy Script checks gates without touching reviewed logic
Four mechanical CI-gate fixes; no production behavior change:

- scripts/_defaults.sh: silence SC2034 for the restart-request out-params
  (AGENTDESK_RESTART_REQUEST_NONCE / _PERSISTENCE_NOT_REQUIRED) that
  deploy-release.sh reads after sourcing; shellcheck can't see cross-file use.
- tests/test_pg_tunnel.py: re-anchor _pg_block() to the new restart-fence
  boundary comment (the old '# #4381: a deploy restarts dcserver' marker was
  replaced by #4735's durable persistence fence). Block content is unchanged.
- runtime_bootstrap/spawns.rs: extract the inline #[cfg(test)] module into a
  sibling spawns_tests.rs (mirrors gateway_lease_tests.rs, #4356) so the file
  drops from 730 to 432 raw LoC, back under the 700 namespace cap. Production
  restart-sentinel / cancellation-guard logic is byte-for-byte unchanged.
- ARCHITECTURE.md: regenerated inventory tree for the new test file.
